chore: refactor AppMarketplace
